Fix convexity detectors not being called
The return value of ConvexityPropagationVisitor.handle_result was
always set to True, which would signal to the main visitor that it was
able to detect convexity information. This resulted in other
extensions not being called.

The Visitor class was returning True when no rules was matched, which
caused the same problem.